+++ classes/prefix_WPgutenberg/blocks/posts/index.js
@@ -425,9 +425,7 @@ export default registerBlockType( 'templates/posts', {
                   <a href="#">
                       {PostImg(postThumb, postTaxonomyFilterOptions, post.id, media[ post.id ])}
                       <div class="post-content">
-                        <h4>
-                          {PostValues(postTextOne, post, postTaxonomyFilterOptions, "link_row1", taxOne)}
-                        </h4>
+                        {PostValues(postTextOne, post, postTaxonomyFilterOptions, "link_row1", taxOne)}
                         {PostValues(postTextTwo, post, postTaxonomyFilterOptions, "link_row2", taxTwo)}
                       </div>
                   </a>